requirements.txt
pygame-ce
pytmx
